+++ Mobile/app/src/main/java/io/github/defolters/homeapp/domain/FormatterValue.java
@@ -2,6 +2,6 @@ package io.github.defolters.homeapp.domain;
 
 public interface FormatterValue {
 
-    String format(int value, int maxValue);
+    String format(float value, int maxValue);
 
 }
